+++ packages/fiori/src/NotificationListItem.ts
@@ -69,29 +69,16 @@ type Footnote = Record<string, any>;
  * <h3>Usage</h3>
  * The component can be used in a standard <code>ui5-list</code>.
  *
- * <h3>CSS Shadow Parts</h3>
- *
- * <ui5-link target="_blank" href="https://developer.mozilla.org/en-US/docs/Web/CSS/::part">CSS Shadow Parts</ui5-link> allow developers to style elements inside the Shadow DOM.
- * <br>
- * The <code>ui5-li-notification</code> exposes the following CSS Shadow Parts:
- * <ul>
- * <li>title-text - Used to style the titleText of the notification list item</li>
- * </ul>
- *
  * <h3>ES6 Module Import</h3>
  *
  * <code>import "@ui5/webcomponents/dist/NotificationListItem.js";</code>
  * <br>
  * <code>import "@ui5/webcomponents/dist/NotificationAction.js";</code> (optional)
  * @constructor
- * @author SAP SE
- * @alias sap.ui.webc.fiori.NotificationListItem
- * @extends sap.ui.webc.fiori.NotificationListItemBase
- * @tagname ui5-li-notification
- * @appenddocs sap.ui.webc.fiori.NotificationAction
+ * @extends NotificationListItemBase
  * @since 1.0.0-rc.8
- * @implements sap.ui.webc.fiori.INotificationListItem, sap.ui.webc.main.IListItem
  * @public
+ * @csspart title-text - Used to style the titleText of the notification list item
  */
 @customElement({
 	tag: "ui5-li-notification",
@@ -116,10 +103,8 @@ class NotificationListItem extends NotificationListItemBase {
 	* <br><br>
 	* <b>Note:</b> by default the <code>titleText</code> and <code>description</code>,
 	* and a <code>ShowMore/Less</code> button would be displayed.
-	* @type {sap.ui.webc.main.types.WrappingType}
-	* @defaultvalue "None"
+	* @default "None"
 	* @public
-	* @name sap.ui.webc.fiori.NotificationListItem.prototype.wrappingType
 	* @since 1.0.0-rc.15
 	*/
 	@property({ type: WrappingType, defaultValue: WrappingType.None })
@@ -150,20 +135,14 @@ class NotificationListItem extends NotificationListItemBase {
 	* we recommend using avatars with 2rem X 2rem in size (32px X 32px). In case you are using the <code>ui5-avatar</code>
 	* you can set its <code>size</code> property to <code>XS</code> to get the required size - <code>&lt;ui5-avatar size="XS">&lt;/ui5-avatar></code>.
 	*
-	* @type {sap.ui.webc.main.IAvatar}
-	* @slot
 	* @public
-	* @name sap.ui.webc.fiori.NotificationListItem.prototype.avatar
 	*/
 	@slot()
 	avatar!: Array<HTMLElement>;
 
 	/**
 	* Defines the elements, displayed in the footer of the of the component.
-	* @type {HTMLElement[]}
-	* @slot
 	* @public
-	* @name sap.ui.webc.fiori.NotificationListItem.prototype.footnotes
 	*/
 	@slot({ type: HTMLElement, individualSlots: true })
 	footnotes!: Array<HTMLElement>;
@@ -175,10 +154,7 @@ class NotificationListItem extends NotificationListItemBase {
 	* <br><br>
 	* <b>Note:</b> Although this slot accepts HTML Elements, it is strongly recommended that you only use text in order to preserve the intended design.
 	*
-	* @type {Node[]}
-	* @slot description
 	* @public
-	* @name sap.ui.webc.fiori.NotificationListItem.prototype.default
 	*/
 	@slot({ type: Node, "default": true })
 	description!: Array<Node>;