+++ packages/main/src/MultiComboBoxGroupItem.ts
@@ -9,13 +9,9 @@ import type { IMultiComboBoxItem } from "./MultiComboBox.js";
  * that can be used to split the <code>ui5-multi-combobox</code> suggestions into groups.
  *
  * @constructor
- * @author SAP SE
- * @alias sap.ui.webc.main.MultiComboBoxGroupItem
- * @extends sap.ui.webc.base.UI5Element
  * @abstract
- * @tagname ui5-mcb-group-item
  * @public
- * @implements sap.ui.webc.main.IMultiComboBoxItem
+ * @implements {IMultiComboBoxItem}
  * @since 1.4.0
  */
 @customElement("ui5-mcb-group-item")
@@ -23,9 +19,7 @@ class MultiComboBoxGroupItem extends UI5Element implements IMultiComboBoxItem {
 	/**
 	 * Defines the text of the component.
 	 *
-	 * @type {string}
-	 * @name sap.ui.webc.main.MultiComboBoxGroupItem.prototype.text
-	 * @defaultvalue ""
+	 * @default ""
 	 * @public
 	 */
 	@property()