+++ packages/main/src/types/TabContainerTabsPlacement.ts
@@ -1,24 +1,18 @@
 /**
  * Tabs placement of TabContainer.
  *
- * @readonly
- * @enum {string}
  * @private
- * @author SAP SE
- * @alias sap.ui.webc.main.types.TabContainerTabsPlacement
  */
 enum TabContainerTabsPlacement {
 	/**
 	 * The tab strip is displayed above the tab content (Default)
 	 * @private
-	 * @type {Top}
 	 */
 	Top = "Top",
 
 	/**
 	 * The tab strip is displayed below the tab content
 	 * @private
-	 * @type {Bottom}
 	 */
 	Bottom = "Bottom",
 }