+++ packages/main/test/specs/ComboBox.spec.js
@@ -1095,4 +1095,72 @@ describe("Keyboard navigation", async () => {
 
 		assert.strictEqual(await input.getProperty("value"), "b", "Value is not autocompleted");
 	});
+
+	it ("Should scroll to items that are in the scroll area upon navigation", async () => {
+		await browser.url(`test/pages/ComboBox.html`);
+		await browser.setWindowSize(1000, 400);
+
+		const combo = await browser.$("#combo-grouping");
+		const input = await combo.shadow$("#ui5-combobox-input");
+		const arrow = await combo.shadow$("[input-icon]");
+
+		await combo.scrollIntoView();
+
+
+		await arrow.click();
+
+		let isInVisibleArea = await browser.executeAsync(async done => {
+			const combobox = document.getElementById("combo-grouping");
+			const picker = await combobox._getPicker();
+			const listItem = picker.querySelector(".ui5-combobox-items-list ui5-li:last-child");
+			const scrollableRect = picker.shadowRoot.querySelector(".ui5-popup-content").getBoundingClientRect();
+			const elementRect = listItem.getBoundingClientRect();
+
+			// Check if the element is within the visible area
+			const isElementAboveViewport = elementRect.bottom < scrollableRect.top;
+			const isElementBelowViewport = elementRect.top > scrollableRect.bottom;
+			const isElementLeftOfViewport = elementRect.right < scrollableRect.left;
+			const isElementRightOfViewport = elementRect.left > scrollableRect.right;
+
+			const isListItemInVisibleArea =  (
+				!isElementAboveViewport &&
+				!isElementBelowViewport &&
+				!isElementLeftOfViewport &&
+				!isElementRightOfViewport
+		);
+
+			done(isListItemInVisibleArea);
+		});
+
+		assert.notOk(isInVisibleArea, "Item should not be displayed in the viewport");
+
+		// click ArrowDown 16 times
+		await input.keys(["ArrowDown", "ArrowDown", "ArrowDown", "ArrowDown", "ArrowDown", "ArrowDown", "ArrowDown", "ArrowDown"]);
+		await input.keys(["ArrowDown", "ArrowDown", "ArrowDown", "ArrowDown", "ArrowDown", "ArrowDown", "ArrowDown", "ArrowDown"]);
+
+		isInVisibleArea = await browser.executeAsync(async done => {
+			const combobox = document.getElementById("combo-grouping");
+			const picker = await combobox._getPicker();
+			const listItem = picker.querySelector(".ui5-combobox-items-list ui5-li:last-child");
+			const scrollableRect = picker.shadowRoot.querySelector(".ui5-popup-content").getBoundingClientRect();
+			const elementRect = listItem.getBoundingClientRect();
+
+			// Check if the element is within the visible area
+			const isElementAboveViewport = elementRect.bottom < scrollableRect.top;
+			const isElementBelowViewport = elementRect.top > scrollableRect.bottom;
+			const isElementLeftOfViewport = elementRect.right < scrollableRect.left;
+			const isElementRightOfViewport = elementRect.left > scrollableRect.right;
+
+			const isListItemInVisibleArea =  (
+				!isElementAboveViewport &&
+				!isElementBelowViewport &&
+				!isElementLeftOfViewport &&
+				!isElementRightOfViewport
+		);
+
+			done(isListItemInVisibleArea);
+		});
+
+		assert.ok(isInVisibleArea, "Item should be displayed in the viewport");
+	});
 });